Former-commit-id: 5a413aa19582782da5e5c784e190843fb252e53b [formerly 4ad86d7db7e5a563aa3039a65dab20bff16cf044]
Former-commit-id: 3f07928360dfa4388c767f53557a7bffaa560903

+++ doc/manual/build.sh
@@ -3,24 +3,29 @@
 if [ $GEM_SET_DEBUG ]; then
     set -x
 fi
+set -e
 
 inkscape -A figures/oq_manual_cover.pdf figures/oq_manual_cover.svg
-pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex &> log.md
-bibtex oq-manual > log.md
-pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex &> log.md
-pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex &> log.md
+
+(pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex
+bibtex oq-manual
+pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex
 makeindex oq-manual.idx
-makeglossaries oq-manual &> log.md
-pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex &> log.md
-cat log.md | egrep "Error|Warning"
-./clean.sh
-if [ "$1" == "--compress" ]; then
-    pdfinfo "oq-manual.pdf" | sed -e 's/^ *//;s/ *$//;s/ \{1,\}/ /g' -e 's/^/  \//' -e '/CreationDate/,$d' -e 's/$/)/' -e 's/: / (/' > .pdfmarks
-    sed -i '1s/^ /[/' .pdfmarks
-    sed -i '/:)$/d' .pdfmarks
-    echo "  /DOCINFO pdfmark" >> .pdfmarks
+makeglossaries oq-manual
+pdflatex -shell-escape -interaction=nonstopmode oq-manual.tex) | egrep -i "error|warning|missing"
 
-    gs -sDEVICE=pdfwrite -dCompatibilityLevel=1.4 -dPDFSETTINGS=/printer -dNOPAUSE -dQUIET -dBATCH -sOutputFile=compressed-oq-manual.pdf oq-manual.pdf .pdfmarks
+if [ -f oq-manual.pdf ]; then
+    ./clean.sh || true
+    if [ "$1" == "--compress" ]; then
+        pdfinfo "oq-manual.pdf" | sed -e 's/^ *//;s/ *$//;s/ \{1,\}/ /g' -e 's/^/  \//' -e '/CreationDate/,$d' -e 's/$/)/' -e 's/: / (/' > .pdfmarks
+        sed -i '1s/^ /[/' .pdfmarks
+        sed -i '/:)$/d' .pdfmarks
+        echo "  /DOCINFO pdfmark" >> .pdfmarks
 
-    mv -f compressed-oq-manual.pdf oq-manual.pdf
+        gs -sDEVICE=pdfwrite -dCompatibilityLevel=1.4 -dPDFSETTINGS=/printer -dNOPAUSE -dQUIET -dBATCH -sOutputFile=compressed-oq-manual.pdf oq-manual.pdf .pdfmarks
+
+        mv -f compressed-oq-manual.pdf oq-manual.pdf
+    fi
+else
+    exit 1
 fi
